Rename Format3Report to Report
